# lib/api-zod/src/generated/types/applicationEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Moving Hiring Pipeline API
 * OpenAPI spec version: 1.0.0
 */
import type { ApplicationEntryStatus } from "./applicationEntryStatus";

export interface ApplicationEntry {
  applicationId: number;
  applicantId: number;
  applicantName: string;
  applicantEmail: string;
  status: ApplicationEntryStatus;
  /** @nullable */
  queuePosition?: number | null;
  appliedAt: Date;
  /** @nullable */
  promotedAt?: Date | null;
  /** @nullable */
  acknowledgeDeadline?: Date | null;
}