# lib/api-zod/src/generated/types/jobQueue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Moving Hiring Pipeline API
 * OpenAPI spec version: 1.0.0
 */
import type { QueueEntry } from "./queueEntry";

export interface JobQueue {
  jobId: number;
  jobTitle: string;
  capacity: number;
  activeCount: number;
  waitlistEntries: QueueEntry[];
}