# lib/api-zod/src/generated/types/pipelineSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Moving Hiring Pipeline API
 * OpenAPI spec version: 1.0.0
 */

export interface PipelineSummary {
  jobId: number;
  jobTitle: string;
  capacity: number;
  activeCount: number;
  waitlistCount: number;
  inactiveCount: number;
  totalApplications: number;
  /** @nullable */
  avgTimeToPromotionSeconds?: number | null;
  /** @nullable */
  avgAcknowledgmentTimeSeconds?: number | null;
  decayEvents: number;
}