# lib/api-zod/src/generated/types/replayEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Moving Hiring Pipeline API
 * OpenAPI spec version: 1.0.0
 */

export interface ReplayEntry {
  applicationId: number;
  applicantId: number;
  applicantName: string;
  status: string;
}